Use try-catch and display a message via DialogCoordinator if failed.
