+++ src/MahApps.Metro.IconPacks.Browser/ExportTemplates/Bitmap.xml
@@ -1,21 +1,11 @@
 ﻿<Canvas xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
+        xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
 		Background="@Background"
         Width="@PageWidth" 
         Height="@PageHeight">
-    <Canvas.LayoutTransform>
-        <MatrixTransform Matrix="@TranformMatrix" />
-    </Canvas.LayoutTransform>
-    <Path Data="@PathData"
-          Stroke="@StrokeColor"
-		  Width="@PageWidth" 
-	      Height="@PageHeight"
-          Fill="@FillColor"
-          Stretch="Uniform" 
-          SnapsToDevicePixels="False"
-          UseLayoutRounding="False"
-          StrokeThickness="@StrokeWidth"
-          StrokeStartLineCap="@StrokeLineCap"
-          StrokeEndLineCap="@StrokeLineCap"
-          StrokeLineJoin="@StrokeLineJoin"/>
+    <iconPacks:PackIcon@IconPackName Kind="@IconKind"
+		                             Width="@PageWidth" 
+	                                 Height="@PageHeight"
+                                     Foreground="@FillColor" />
 </Canvas>
