+++ src/MahApps.Metro.IconPacks.Browser/ExportTemplates/WPF.xml
@@ -26,7 +26,9 @@ Iconpack-License:	@IconPackLicense
     <Path Data="@PathData"
           Stroke="@StrokeColor"
           Fill="@FillColor"
-          Stretch="Uniform" 
+          Stretch="Uniform"
+          Width="@PageWidth"
+          Height="@PageHeight"
           SnapsToDevicePixels="False"
           UseLayoutRounding="False"
           StrokeThickness="@StrokeWidth"
